refactor(firewood): Move to `vms/evm`
