+++ graft/coreth/core/BUILD.bazel
@@ -51,12 +51,12 @@ go_library(
         "//graft/coreth/precompile/modules",
         "//graft/coreth/precompile/precompileconfig",
         "//graft/evm/core/state/snapshot",
-        "//graft/evm/firewood",
         "//graft/evm/triedb/hashdb",
         "//graft/evm/triedb/pathdb",
         "//utils/set",
         "//vms/evm/acp176",
         "//vms/evm/acp226",
+        "//vms/evm/firewood",
         "//vms/evm/predicate",
         "//vms/evm/sync/customrawdb",
         "@com_github_ava_labs_firewood_go_ethhash_ffi//:ffi",
@@ -132,7 +132,6 @@ graft_go_test(
         "//graft/coreth/precompile/contracts/warp",
         "//graft/coreth/precompile/precompileconfig",
         "//graft/evm/core/state/pruner",
-        "//graft/evm/firewood",
         "//graft/evm/triedb/pathdb",
         "//snow/engine/snowman/block",
         "//upgrade",
@@ -140,6 +139,7 @@ graft_go_test(
         "//utils",
         "//utils/set",
         "//vms/evm/acp176",
+        "//vms/evm/firewood",
         "//vms/evm/predicate",
         "//vms/evm/sync/customrawdb",
         "@com_github_ava_labs_libevm//accounts/abi",
