+++ graft/coreth/eth/BUILD.bazel
@@ -34,10 +34,10 @@ go_library(
         "//graft/coreth/node",
         "//graft/coreth/params",
         "//graft/evm/core/state/pruner",
-        "//graft/evm/firewood",
         "//graft/evm/rpc",
         "//utils/timer/mockable",
         "//version",
+        "//vms/evm/firewood",
         "//vms/evm/sync/customrawdb",
         "@com_github_ava_labs_firewood_go_ethhash_ffi//:ffi",
         "@com_github_ava_labs_libevm//accounts",
