+++ graft/coreth/eth/tracers/BUILD.bazel
@@ -19,8 +19,8 @@ go_library(
         "//graft/coreth/core",
         "//graft/coreth/internal/ethapi",
         "//graft/coreth/params",
-        "//graft/evm/firewood",
         "//graft/evm/rpc",
+        "//vms/evm/firewood",
         "@com_github_ava_labs_libevm//common",
         "@com_github_ava_labs_libevm//common/hexutil",
         "@com_github_ava_labs_libevm//core/state",
