+++ graft/coreth/internal/ethapi/BUILD.bazel
@@ -25,8 +25,8 @@ go_library(
         "//graft/coreth/eth/gasestimator",
         "//graft/coreth/params",
         "//graft/coreth/plugin/evm/customtypes",
-        "//graft/evm/firewood",
         "//graft/evm/rpc",
+        "//vms/evm/firewood",
         "@com_github_ava_labs_libevm//accounts",
         "@com_github_ava_labs_libevm//accounts/abi",
         "@com_github_ava_labs_libevm//accounts/keystore",
