+++ graft/coreth/plugin/evm/BUILD.bazel
@@ -57,7 +57,6 @@ go_library(
         "//graft/coreth/precompile/registry",
         "//graft/coreth/warp",
         "//graft/evm/constants",
-        "//graft/evm/firewood",
         "//graft/evm/message",
         "//graft/evm/rpc",
         "//graft/evm/sync/client",
@@ -89,6 +88,7 @@ go_library(
         "//vms/evm/acp176",
         "//vms/evm/acp226",
         "//vms/evm/database",
+        "//vms/evm/firewood",
         "//vms/evm/metrics/prometheus",
         "//vms/evm/sync/customrawdb",
         "@com_github_ava_labs_firewood_go_ethhash_ffi//:ffi",
