+++ graft/coreth/plugin/evm/atomic/vm/BUILD.bazel
@@ -106,7 +106,6 @@ graft_go_test(
         "//graft/coreth/plugin/evm/upgrade/ap0",
         "//graft/coreth/plugin/evm/upgrade/ap1",
         "//graft/coreth/plugin/evm/vmtest",
-        "//graft/evm/firewood",
         "//graft/evm/utils",
         "//graft/evm/utils/utilstest",
         "//ids",
@@ -130,6 +129,7 @@ graft_go_test(
         "//utils/units",
         "//vms/components/avax",
         "//vms/components/chain",
+        "//vms/evm/firewood",
         "//vms/evm/predicate",
         "//vms/evm/sync/customrawdb",
         "//vms/secp256k1fx",
