+++ graft/coreth/plugin/evm/vmtest/BUILD.bazel
@@ -28,7 +28,6 @@ go_library(
         "//graft/coreth/plugin/evm/extension",
         "//graft/coreth/plugin/evm/upgrade/ap3",
         "//graft/evm/constants",
-        "//graft/evm/firewood",
         "//graft/evm/sync/client",
         "//graft/evm/sync/engine",
         "//graft/evm/sync/synctest",
@@ -47,6 +46,7 @@ go_library(
         "//utils/set",
         "//vms/components/chain",
         "//vms/evm/database",
+        "//vms/evm/firewood",
         "//vms/evm/predicate",
         "//vms/evm/sync/customrawdb",
         "@com_github_ava_labs_libevm//common",
