+++ graft/evm/sync/engine/BUILD.bazel
@@ -17,7 +17,6 @@ go_library(
         "//database/merkle/firewood/syncer",
         "//database/versiondb",
         "//graft/evm/core/state/snapshot",
-        "//graft/evm/firewood",
         "//graft/evm/message",
         "//graft/evm/sync/block",
         "//graft/evm/sync/client",
@@ -29,6 +28,7 @@ go_library(
         "//snow",
         "//snow/engine/snowman/block",
         "//vms/components/chain",
+        "//vms/evm/firewood",
         "@com_github_ava_labs_libevm//common",
         "@com_github_ava_labs_libevm//core/types",
         "@com_github_ava_labs_libevm//ethdb",
