+++ graft/evm/sync/evmstate/BUILD.bazel
@@ -59,7 +59,6 @@ graft_go_test(
         "//codec",
         "//database/merkle/firewood/syncer",
         "//graft/evm/core/state/snapshot",
-        "//graft/evm/firewood",
         "//graft/evm/message",
         "//graft/evm/message/messagetest",
         "//graft/evm/sync/client",
@@ -70,6 +69,7 @@ graft_go_test(
         "//graft/evm/utils/utilstest",
         "//ids",
         "//network/p2p/p2ptest",
+        "//vms/evm/firewood",
         "//vms/evm/sync/customrawdb",
         "@com_github_ava_labs_firewood_go_ethhash_ffi//:ffi",
         "@com_github_ava_labs_libevm//common",
