+++ graft/subnet-evm/core/BUILD.bazel
@@ -41,7 +41,6 @@ go_library(
         "//database",
         "//graft/evm/constants",
         "//graft/evm/core/state/snapshot",
-        "//graft/evm/firewood",
         "//graft/evm/triedb/hashdb",
         "//graft/evm/triedb/pathdb",
         "//graft/subnet-evm/commontype",
@@ -63,6 +62,7 @@ go_library(
         "//graft/subnet-evm/stateupgrade",
         "//utils/set",
         "//vms/evm/acp226",
+        "//vms/evm/firewood",
         "//vms/evm/predicate",
         "//vms/evm/sync/customrawdb",
         "@com_github_ava_labs_firewood_go_ethhash_ffi//:ffi",
@@ -126,7 +126,6 @@ graft_go_test(
     ],
     deps = [
         "//graft/evm/core/state/pruner",
-        "//graft/evm/firewood",
         "//graft/evm/triedb/pathdb",
         "//graft/subnet-evm/commontype",
         "//graft/subnet-evm/consensus",
@@ -150,6 +149,7 @@ graft_go_test(
         "//upgrade/upgradetest",
         "//utils",
         "//utils/set",
+        "//vms/evm/firewood",
         "//vms/evm/predicate",
         "//vms/evm/sync/customrawdb",
         "@com_github_ava_labs_libevm//accounts/abi",
