+++ graft/subnet-evm/eth/BUILD.bazel
@@ -21,7 +21,6 @@ go_library(
     importpath = "github.com/ava-labs/avalanchego/graft/subnet-evm/eth",
     deps = [
         "//graft/evm/core/state/pruner",
-        "//graft/evm/firewood",
         "//graft/evm/rpc",
         "//graft/subnet-evm/commontype",
         "//graft/subnet-evm/consensus",
@@ -39,6 +38,7 @@ go_library(
         "//graft/subnet-evm/params",
         "//utils/timer/mockable",
         "//version",
+        "//vms/evm/firewood",
         "//vms/evm/sync/customrawdb",
         "@com_github_ava_labs_firewood_go_ethhash_ffi//:ffi",
         "@com_github_ava_labs_libevm//accounts",
