+++ graft/subnet-evm/plugin/evm/BUILD.bazel
@@ -40,7 +40,6 @@ go_library(
         "//database/prefixdb",
         "//database/versiondb",
         "//graft/evm/constants",
-        "//graft/evm/firewood",
         "//graft/evm/message",
         "//graft/evm/rpc",
         "//graft/evm/sync/client",
@@ -94,6 +93,7 @@ go_library(
         "//vms/components/chain",
         "//vms/evm/acp226",
         "//vms/evm/database",
+        "//vms/evm/firewood",
         "//vms/evm/metrics/prometheus",
         "//vms/evm/sync/customrawdb",
         "//vms/evm/uptimetracker",
@@ -143,7 +143,6 @@ graft_go_test(
         "//database/memdb",
         "//database/prefixdb",
         "//graft/evm/constants",
-        "//graft/evm/firewood",
         "//graft/evm/rpc",
         "//graft/evm/sync/client",
         "//graft/evm/sync/engine",
@@ -206,6 +205,7 @@ graft_go_test(
         "//vms/evm/acp176",
         "//vms/evm/acp226",
         "//vms/evm/database",
+        "//vms/evm/firewood",
         "//vms/evm/predicate",
         "//vms/evm/sync/customrawdb",
         "//vms/evm/uptimetracker",
