+++ graft/subnet-evm/tests/BUILD.bazel
@@ -15,9 +15,9 @@ go_library(
     importpath = "github.com/ava-labs/avalanchego/graft/subnet-evm/tests",
     deps = [
         "//graft/evm/core/state/snapshot",
-        "//graft/evm/firewood",
         "//graft/evm/triedb/hashdb",
         "//graft/evm/triedb/pathdb",
+        "//vms/evm/firewood",
         "//vms/evm/sync/customrawdb",
         "@com_github_ava_labs_libevm//common",
         "@com_github_ava_labs_libevm//core/rawdb",
