+++ vms/evm/firewood/hash_test.go
